Extracted events made ready fire once
